+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Rename run command to gen
 * Replace -r run ID with -i gen ID
 * Change default out file name from 'out' to 'data'
+* Add template params: integer, float, date, select, word, first_name, and last_name
 
 ### 0.0.1
 * Initial version